requirements.txt
transformers==4.22.0
torch==1.11.0
datasets==2.4.0
Pillow==9.0.1
Flask==2.0.
Werkzeug==2.1.0
